# openapi.yaml
openapi: 3.0.3
info:
  title: Omniscience API
  description: REST API for managing educational data including courses, classes, students, and attendance records
  version: 1.0.0
  contact:
    name: Omniscience API Team
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Supabase authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...
    Grade:
      type: object
      properties:
        id:
          type: integer
          format: uint
        evaluation_item_id:
          type: integer
          format: uint
        registration_id:
          type: integer
          format: uint
        score:
          type: number
          format: float
        percentage:
          type: number
          format: float
          description: Automatically calculated percentage score
        letter_grade:
          type: string
          description: Automatically assigned letter grade
        is_excused:
          type: boolean
        is_late:
          type: boolean
        late_penalty:
          type: number
          format: float
        comments:
          type: string
        graded_by:
          type: string
        graded_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [DRAFT, PUBLISHED]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - evaluation_item_id
        - registration_id
...
    CategoryScore:
      type: object
      properties:
        category_id:
          type: integer
          format: uint
        category_name:
          type: string
        weight:
          type: number
          format: float
        score:
          type: number
          format: float
        grades:
          type: array
          items:
            $ref: '#/components/schemas/Grade'
    StudentReport:
      type: object
      properties:
        registration_id:
          type: integer
          format: uint
        student_class_id:
          type: integer
          format: uint
        category_scores:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/CategoryScore'
        total_grades:
          type: integer
        final_grade:
          $ref: '#/components/schemas/FinalGrade'
    EvaluationStats:
      type: object
      properties:
        evaluation_item_id:
          type: integer
          format: uint
        evaluation_name:
          type: string
        category_name:
          type: string
        total_graded:
          type: integer
        average_percentage:
          type: number
          format: float
        min_percentage:
          type: number
          format: float
        max_percentage:
          type: number
          format: float
        passing_count:
          type: integer
        failing_count:
          type: integer
    ClassReport:
      type: object
      properties:
        student_class_id:
          type: integer
          format: uint
        class_name:
          type: string
        course_name:
          type: string
        evaluation_stats:
          type: array
          items:
            $ref: '#/components/schemas/EvaluationStats'
        total_students:
          type: integer
        passing_students:
          type: integer
        average_final_grade:
          type: number
          format: float
        final_grades:
          type: array
          items:
            $ref: '#/components/schemas/FinalGrade'
    FinalGrade:
      type: object
      properties:
        id:
          type: integer
          format: uint
        registration_id:
          type: integer
          format: uint
        student_class_id:
          type: integer
          format: uint
        calculated_percentage:
          type: number
          format: float
        letter_grade:
          type: string
        is_passing:
          type: boolean
        category_scores:
          type: object
          additionalProperties:
            type: object
            properties:
              category_id:
                type: integer
                format: uint
              score:
                type: number
                format: float
              weight:
                type: number
                format: float
        status:
          type: string
          enum: [DRAFT, PUBLISHED]
        calculation_date:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CalculateFinalGradesResponse:
      type: object
      properties:
        message:
          type: string
        student_class_id:
          type: integer
          format: uint
security:
  - bearerAuth: []
paths:
...
  /reports/student/{registration_id}:
    get:
      summary: Get student report
      description: Get detailed grade report for a student including all published grades grouped by category
      operationId: getStudentReport
      tags:
        - Reports
      security:
        - bearerAuth: []
      parameters:
        - name: registration_id
          in: path
          description: Registration ID
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentReport'
        '404':
          description: Registration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /reports/class/{student_class_id}:
    get:
      summary: Get class report
      description: Get performance statistics and analytics for an entire class
      operationId: getClassReport
      tags:
        - Reports
      security:
        - bearerAuth: []
      parameters:
        - name: student_class_id
          in: path
          description: Student Class ID
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassReport'
        '404':
          description: Class not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /final-grades/{registration_id}:
    get:
      summary: Get final grade
      description: Get calculated final grade for a student
      operationId: getFinalGrade
      tags:
        - Final Grades
      security:
        - bearerAuth: []
      parameters:
        - name: registration_id
          in: path
          description: Registration ID
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalGrade'
        '404':
          description: Final grade not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /final-grades/calculate/{student_class_id}:
    post:
      summary: Calculate final grades
      description: Trigger recalculation of final grades for all students in a class
      operationId: calculateFinalGrades
      tags:
        - Final Grades
      security:
        - bearerAuth: []
      parameters:
        - name: student_class_id
          in: path
          description: Student Class ID
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: Final grades calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculateFinalGradesResponse'
        '404':
          description: Class not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...